python_requirements.txt
matplotlib==3.8.2
numpy==1.26.2
pandas==2.1.3
protobuf==4.25.1
python-dotenv==1.0.0
pytz==2023.3.post1
scikit-learn==1.3.2
seaborn==0.13.0
statsmodels==0.14.0
tornado # for webagg on optuna viz (not necessary for non-GUI env)
SQLAlchemy==2.0.23
httpx==0.25.2
psycopg2-binary==2.9.9
dagster==1.6.3 
dagster-cloud==1.6.3 
dagster-webserver==1.6.3